Move `useDoc` hook to its own file in `@/hooks`
