+++ src/hooks/use-doc.ts
@@ -1,0 +1,5 @@
+import { useContext } from 'react';
+
+import { DocContext } from '@/components/layout/provider/doc';
+
+export const useDoc = () => useContext(DocContext);
